Good work.
Main commit is a functioning richText widget. There were some issues with IE/Firefox compatibility, but I think those are sorted.

Change #2 is public tags. Right now you need to set them manually in the DB. 

Small changes to topic & tag widgets that shoudl prevent duplication problems from saving, then saving again, but not having had a load() in there.

git-svn-id: file:///a/svnrepos/hippo/code/trunk/HippoTest@98 b8d5060c-ad15-0410-9d10-96fc93d4f2a4

+++ src/com/aavu/server/service/impl/TagServiceImpl.java
@@ -5,12 +5,13 @@ import java.util.List;
 import org.apache.log4j.Logger;
 
 import com.aavu.client.domain.Tag;
+import com.aavu.client.exception.PermissionDeniedException;
 import com.aavu.server.dao.TagDAO;
 import com.aavu.server.service.UserService;
 
 public class TagServiceImpl implements com.aavu.server.service.TagService {
 	private static final Logger log = Logger.getLogger(TagServiceImpl.class);
-	
+
 	private TagDAO tagDAO;
 	private UserService userService;
 
@@ -33,7 +34,7 @@ public class TagServiceImpl implements com.aavu.server.service.TagService {
 
 	public Tag getTagAddIfNew(String tagName) {
 		log.debug("load tag named: "+tagName);
-		
+
 		Tag rt = tagDAO.getTag(userService.getCurrentUser(),tagName);
 		if(null == rt){
 			log.debug("was null, creating ");
@@ -42,12 +43,12 @@ public class TagServiceImpl implements com.aavu.server.service.TagService {
 			t.setPublicVisible(false);
 			t.setUser(userService.getCurrentUser());
 			tagDAO.save(t);
-			
+
 			log.debug("created: "+t.getId());
 			return t;
 		}else{
 			log.debug("existed "+rt.getId()+"\n"+rt.toPrettyString());
-			
+
 			return rt;
 		}
 	}
@@ -62,19 +63,24 @@ public class TagServiceImpl implements com.aavu.server.service.TagService {
 
 
 
-	public void removeTag(Tag selectedTag) {
-		tagDAO.removeTag(userService.getCurrentUser(),selectedTag);
+	public void removeTag(Tag selectedTag) throws PermissionDeniedException {
+		if(userService.getCurrentUser().equals(selectedTag.getUser())){
+			tagDAO.removeTag(userService.getCurrentUser(),selectedTag);
+		}else{
+			log.warn("User: "+userService.getCurrentUser().getUsername()+" tried to delete tag of "+selectedTag.getUser().getUsername());
+			throw new PermissionDeniedException("No permission to delete tag");
+		}
 	}
 
 
 
 
 	public void save(Tag selectedTag) {
-		
+
 		log.debug("Servic tag.save() setting user");
-		
+
 		selectedTag.setUser(userService.getCurrentUser());
-		
+
 		tagDAO.save(selectedTag);
 	}
